+++ security_tests_m8/suites/__init__.py
@@ -7,6 +7,7 @@ from security_tests_m8.suites.algorithms import (
     HS256WeakKeySuite,
     JWKSSuite,
 )
+from security_tests_m8.suites.api_key import ApiKeyRedisDegradedSuite
 from security_tests_m8.suites.deployment import DeploymentPreflightSuite
 from security_tests_m8.suites.media import MediaInternalExposureSuite
 from security_tests_m8.suites.service import (
@@ -38,6 +39,7 @@ from security_tests_m8.suites.universal import (
 )
 
 __all__ = [
+    "ApiKeyRedisDegradedSuite",
     "ApiKeySuite",
     "AsymmetricJWTSuite",
     "AuthAttackSuite",